Change  field int cast to float cast

+++ PLATER/services/util/graph_adapter.py
@@ -117,7 +117,7 @@ class GraphInterface:
             hits = self.convert_to_dict(result)
             for hit in hits:
                 hit["node"] = dict(dict(hit["node"])["properties"])
-                hit["score"] = int(hit["score"])
+                hit["score"] = float(hit["score"])
             hits.sort(key=lambda hit: hit["score"], reverse=True)
             return hits
